Minimizing and equalizing torques.

+++ calculations/transformations.py
@@ -102,7 +102,7 @@ def R_B1(qb, q1):
         [math.cos(q1 + qb), -math.sin(q1 + qb), 0],
         [math.sin(q1 + qb), math.cos(q1 + qb), 0],
         [0, 0, 1]
-    ])
+    ], dtype = np.float32)
 
 def R_12(q2):
     """Rotation matrix from 1st to 2nd leg-segment.
@@ -117,7 +117,7 @@ def R_12(q2):
         [math.cos(q2), -math.sin(q2), 0],
         [0, 0, -1],
         [math.sin(q2), math.cos(q2), 0]
-    ])
+    ], dtype = np.float32)
 
 def R_23(q3):
     """Rotation matrix from 2nd to 3rd leg-segment.
@@ -132,7 +132,7 @@ def R_23(q3):
         [math.cos(q3), -math.sin(q3), 0],
         [math.sin(q3), math.cos(q3), 0],
         [0, 0, 1]
-    ])
+    ], dtype = np.float32)
 
 def R_B2(qb, q1, q2):
     """Rotation matrix from spider's to 2nd segment's origin.
@@ -149,11 +149,11 @@ def R_B2(qb, q1, q2):
         [math.cos(q2) * math.cos(q1 + qb), -math.cos(q1 + qb) * math.sin(q2), math.sin(q1 + qb)],
         [math.cos(q2) * math.sin(q1 + qb), -math.sin(q2) * math.sin(q1 + qb), -math.cos(q1 + qb)],
         [math.sin(q2), math.cos(q2), 0]
-    ])
+    ], dtype = np.float32)
 
 def R_B3(qb, q1, q2, q3):
     return np.array([
         [math.cos(q2 + q3) * math.cos(q1 + qb), -math.cos(q1 + qb) * math.sin(q2 + q3), math.sin(q1 + qb)],
         [math.cos(q2 + q3) * math.sin(q1 + qb), -math.sin(q2 + q3) * math.sin(q1 + qb), -math.cos(q1 + qb)],
         [math.sin(q2 + q3), math.cos(q2 + q3), 0],
-    ])
+    ], dtype = np.float32)
